requirements.txt
fastapi[standard]
uvicorn[standard]
sqlmodel
pwdlib[argon2]
PyJWT
psycopg2-binary
python-dotenv
httpx
pytest
pytest-asyncio